requirements.txt
Pillow
pycryptodome
stepic
cryptography
pytest
